Split block.py 2/8: extract branching subsystem → models/branching.py
Part 2/8, on the block_base foundation. branching.py is a leaf module (lazy call-time imports);
ConditionalBlock stays in block.py and subclasses Block (now re-exported from block_base).
